barbearia_borba/requirements.txt
Django>=3.2
djangorestframework>=3.12
djangorestframework-simplejwt>=4.0
psycopg2>=2.9
celery>=5.1
redis>=4.0
twilio
